+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased
+### Added
+- `dsnpStartBlockNumber` to config to allow setConfig to default a different value than `0` for `fromBlock`
+
 ### Changed
 - Removed esmodule build. All imports will fallback to esmodule compatible commonjs modules
 - Better cleaner build (removed multimodule script need, dist package directory, and better package.json)